Add explicit conversions

+++ rcl_lifecycle/src/com_interface.c
@@ -344,8 +344,8 @@ rcl_lifecycle_com_interface_publish_notification(
   }
 
   com_interface->msg.stamp.sec = (int32_t) RCL_NS_TO_S(timestamp);
-  com_interface->msg.stamp.nanosec = (timestamp % RCL_S_TO_NS(1));
-  com_interface->msg.transition.id = transition->id;
+  com_interface->msg.stamp.nanosec = (uint32_t) (timestamp % RCL_S_TO_NS(1));
+  com_interface->msg.transition.id = (uint8_t) transition->id;
   rosidl_runtime_c__String__assign(&com_interface->msg.transition.label, transition->label);
   com_interface->msg.start_state.id = transition->start->id;
   rosidl_runtime_c__String__assign(&com_interface->msg.start_state.label, transition->start->label);
